Fixed RS project creation
The package explorer filter now has an entry for gradle build folders.
The filters are run when the project is created via resource change
events. We were creating folders without setting them as local
(whatever that means in eclipse-land) and the gradle build filter was
unable to deal with the folders.